+++ tests/test_ops/test_image/test_transforms.py
